Refactor : 에러코드 분리 및 기존 에러코드 경로 삭제
